requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
graphene-django>=3.0
django-filter>=24.2
gunicorn
